% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_of_vehicle_stolen_age.R
\name{age_of_vehicle_stolen_age}
\alias{age_of_vehicle_stolen_age}
\title{age_of_vehicle_stolen_age}
\usage{
age_of_vehicle_stolen_age(config, master_config, subtable)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
