% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_significance_column.R
\name{append_significance_column}
\alias{append_significance_column}
\title{append_significance_column}
\usage{
append_significance_column(
  dataframe,
  column_1,
  column_2,
  design_1 = 1.2,
  design_2 = 1.2,
  row_of_unweighted_bases,
  threshold = 1.96,
  significance_indicator = "*",
  insignificance_indicator = " "
)
}
\arguments{
\item{dataframe}{A numeric \code{dataframe} containing 1 row of unweighted bases and at least 2 columns.}

\item{column_1}{An integer specifying a column of \code{dataframe}, different to \code{column_2}.}

\item{column_2}{An integer specifying a column of \code{dataframe}, different to \code{column_1}.}

\item{design_1}{TO BE ADDED LATER.}

\item{design_2}{TO BE ADDED LATER.}

\item{row_of_unweighted_bases}{An integer indicating which row of \code{dataframe} holds the unweighted bases.}

\item{threshold}{TO BE ADDED LATER.}

\item{significance_indicator}{A string specifying how significance should be indicated in the appended column.}

\item{insignificance_indicator}{A string specifying how insignificance should be indicated in the appended column.}
}
\value{
The original \code{dataframe}, except with the new significance column appended to its right.
}
\description{
Performs z-tests across two columns of a dataframe, testing each cell of the first column against the
corresponding cell of the second; the results of these tests are appended as a column to the right of the original
dataframe.
}
\details{
This function repeatedly calls \code{autotable::z_test_crime} to test each value of \code{column_1} against
its counterpart in \code{column_2}. If a result is significanct, it will be indicated as such by
\code{significance_indicator} in the appended column; if it is not, then it will be indicated as such by
\code{insignificance_indicator}.
}
