% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_weight.R
\name{apply_weight}
\alias{apply_weight}
\title{apply_weight}
\usage{
apply_weight(df, columns, weight)
}
\arguments{
\item{df}{The dataframe to manuplicate}

\item{columns}{The columns to apply the weight too}

\item{weight}{The column by which to mutiple the other columns by, supplied as a character string}
}
\value{
The original df but witht the selected columns altered
}
\description{
applies a weight: mutiplies selected columns by another column (the weight)
}
\details{
The function takes a list of columns and mutiplies each value (or row) by the value in another column
which is the weight column
}
\examples{
apply_weight(mtcars, columns= c("hp", "drat"), weight= "carb")

}
