% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_damage.R
\name{cost_of_damage}
\alias{cost_of_damage}
\title{cost_of_damage}
\usage{
cost_of_damage(config, master_config, return_as_dataframes = FALSE)
}
\arguments{
\item{config}{A list of lists (and other items) which is the result of applying \code{yaml::read.yaml()} to an appropriate "Nature
of Crime" .yaml file, and then \code{autotable::process_config()}.}

\item{master_config}{A list of lists (and other objects) which is the result of applying \code{yaml::read.yaml()} to the master
configuration file.}
}
\value{
The "Cost of damage" table for the headline crime type specified by \code{config}, as a \code{GPTable}
object.
}
\description{
Builds the "Cost of damage" table for a headline crime type, as seen in the "Nature of Crime" dataset for that
headline crime type.
}
\details{
This function builds a "Cost of damage" table based on two configuration files. The first config file, \code{config}
corresponds to one of the headline crime types, and chiefly determines the output of the table. The second, \code{master_config},
allows you to tweak things that apply to multiple tables (probably including this one), with the idea being that it's easier to
change the master config file once than it is to update each config file for those multiple tables - which could potentially
constitute changing 50 different files... bothersome and time-consuming).

In this function, each subtable of the table is processed in turn. The processing of the table is handled by lower-order functions,
one corresponding to each part of the table. For the "Cost of damage" table, there is currently only 1 part, which is built by the
function \code{cost_of_damage_cost}, which returns a data frame.

The function then processes this data frame into a \code{GPTable} object (see the User Guide for more details) which is what is
returned.
}
\examples{
config = autotable::process_config(yaml::read.yaml("configs/cost_of_damage/BURGLARY.yaml"))
master_config = yaml::read_yaml(file = "configs/MASTER.yaml")
cost_of_damage(config = config, master_config = master_config)

}
\seealso{
See the following URL for an example of a Nature of Crime dataset (where the headline crime type in the example is
"Vehicle-related theft"): \url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/natureofcrimevehiclerelatedtheft}
}
\author{
Eleanor Scott-Allen, Katie Edser, Matthew Morrison, Nicholas O'Neil; ONS Centre for Crime and Justice, Titchfield, PO15 5RR.
}
