% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_of_damage_cost.R
\name{cost_of_damage_cost}
\alias{cost_of_damage_cost}
\title{cost_of_damage_cost}
\usage{
cost_of_damage_cost(config, master_config, subtable)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
