% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotional_impact_affected_or_not.R
\name{emotional_impact_affected_or_not}
\alias{emotional_impact_affected_or_not}
\title{emotional_impact_affected_or_not}
\usage{
emotional_impact_affected_or_not(config, master_config, subtable)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
