% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emotional_impact_extent_of_impact.R
\name{emotional_impact_extent_of_impact}
\alias{emotional_impact_extent_of_impact}
\title{emotional_impact_extent_of_impact}
\usage{
emotional_impact_extent_of_impact(config, master_config, subtable)
}
\arguments{
\item{config}{A string corresponding to the path of a YAML file, where this YAML serves as the configuration file for the
function. This YAML must exhibit a particular structure/format, since the function accesses certain levels and fields of it
by name. For more information on the format required, please see the User Guide (found in the "resources" folder of the
project) or inspect a valid YAML file from previous successful runs.}

\item{master_config}{To do.}

\item{subtable}{An integer indicating which subtable to consider; if we denote the value of \code{subtable} by \code{n},
then the subtable considered will be that corresponding to the \code{n}th item in
\code{config$ALL_PARTS$OFFENCE_CODES}.}
}
\value{
Returns a dataframe with the requisite percentages and unweighted bases for the EoI part of the specified subtable.
}
\description{
Builds the "Percentage emotionally affected" (EoI) part/half of one subtable of an "Emotional impact of
incidents" Table for a Nature of Crime dataset.
}
\details{
This function builds the EoI part of one subtable (specified by \code{subtable}) based largely on a configuration
file specified by \code{config}.
}
\examples{
emotional_impact_extent_of_impact(config = "configs/emotional_impact/CRIMINAL_DAMAGE.yaml", data_directory =
"data/backseries/", subtable = 2)

}
