% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sheet_names.R
\name{generate_sheet_names}
\alias{generate_sheet_names}
\title{generate_sheet_names}
\usage{
generate_sheet_names(Workbook_structure)
}
\arguments{
\item{Workbook_structure}{A numeric vector, where each slot in the vector respresents a different main table, and the number
in that slot represents how many subtables that table has. For instance, the vector \code{c(3,4,6,8)} indicates that the first
table has 8 subtables, the second table has 4, the third has 6, and the fourth has 8. Each element of the vector should be
an integer between 1 and 25 inclusive.}
}
\value{
A character vector of the sheet names, in order.
}
\description{
Appropriately generates sheet names for an Excel workbook-to-be, where each sheet name is of the form "Table XY",
for X a number, and Y a lowercase letter of the alphabet or an empty string.
}
\details{
This function takes in a numeric vector \code{Workbook_structure} from which it generates the sheet names, which are
returned as a character vector. The function supports a maximum of 25 subtables per table.
}
\examples{
\code{generate_sheet_names(Workbook_structure = c(1, 3, 2))} # Returns c("Table 1", "Table 2a", "Table 2b",
"Table 2c", "Table 3a", "Table 3b")

}
