% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_sig_test.R
\name{mean_sig_test}
\alias{mean_sig_test}
\title{mean_sig_test}
\usage{
mean_sig_test(
  estimate_1,
  estimate_2,
  base_1,
  base_2,
  sd_1,
  sd_2,
  design = 1.2,
  threshold = 1.96,
  significance_indicator = 1,
  insignificance_indicator = 0
)
}
\arguments{
\item{estimate_1}{Documentation to be added later.}

\item{estimate_2}{Documentation to be added later.}

\item{base_1}{Documentation to be added later.}

\item{base_2}{Documentation to be added later.}

\item{threshold}{Documentation to be added later.}

\item{significance_indicator}{Documentation to be added later.}

\item{insignificance_indicator}{Documentation to be added later.}

\item{design_1}{Documentation to be added later.}

\item{design_2}{Documentation to be added later.}
}
\value{
Returns the back-series as a data frame of percentages with the unweighted base as the last row.
}
\description{
Documentation to be added later.
}
\details{
Documentation to be added later.
}
\examples{

}
