% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_incidents_against_men_and_women.R
\name{number_of_incidents_against_men_and_women}
\alias{number_of_incidents_against_men_and_women}
\title{number_of_incidents_against_men_and_women}
\usage{
number_of_incidents_against_men_and_women(
  config,
  master_config,
  return_as_dataframes = FALSE
)
}
\arguments{
\item{config}{Documentation to be done later.}

\item{master-config}{Documentation to be done later.}
}
\description{
Documentation to be done later.
}
\details{
Documentation to be done later.
}
