% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_of_incidents_build_sex_percentages.R
\name{number_of_incidents_build_sex_percentages}
\alias{number_of_incidents_build_sex_percentages}
\title{number_of_incidents_build_sex_percentages}
\usage{
number_of_incidents_build_sex_percentages(config, master_config, subtable)
}
\arguments{
\item{config}{Documentation to be done later.}

\item{master-config}{Documentation to be done later.}
}
\description{
Documentation to be done later.
}
\details{
Documentation to be done later.
}
