% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/number_of_offenders_under_the_influence_proportions.R
\name{number_of_offenders_under_the_influence_proportions}
\alias{number_of_offenders_under_the_influence_proportions}
\title{number_of_offenders_under_the_influence_proportions}
\usage{
number_of_offenders_under_the_influence_proportions(
  config,
  master_config,
  subtable
)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
