% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offender_characteristics.R
\name{offender_characteristics}
\alias{offender_characteristics}
\title{offender_characteristics}
\usage{
offender_characteristics(config, master_config, return_as_dataframes = FALSE)
}
\arguments{
\item{config}{A list of lists (and other items) which is the result of applying \code{yaml::read.yaml()} to an appropriate "Nature
of Crime" .yaml file, and then \code{autotable::process_config()}.}

\item{master_config}{A list of lists (and other items) which is the result of applying \code{yaml::read.yaml()} to the master
configuration file.}
}
\value{
The "Offender characteristics" table for the headline crime type specified by \code{config}, as a \code{GPTable}
object.
}
\description{
Builds the "Offender characteristics" table for a headline crime type, as seen in the "Nature of Crime" dataset
for that headline crime type.
}
\details{
This function builds an "Offender characteristics" table based on two configuration files. The first config
file, \code{config}, corresponds to one of the headline crime types, and chiefly determines the output of the table. The second,
\code{master_config}, allows you to tweak things that apply to multiple tables (probably including this one), with the idea being
that it's easier to change the master config file once than it is to update each config file for those multiple tables - which
could potentially constitute changing 50 different files... bothersome and time-consuming).

In this function, each subtable of the table is processed in turn. The processing of the table is handled by lower-order functions,
one corresponding to each part of the table. For the "Offender characteristics" table, there are currently 5 parts:
"Was the respondent able to say something about the offender?", "Number of offenders", "Sex of offenders", "Age of offenders", and
"Relationship to victim"; these are built by the functions \code{offender_characteristics_able_to_say_something},
\code{offender_characteristics_number_of_offenders}, \code{offender_characteristics_sex_of_offenders},
\code{offender_characteristics_age_of_offenders} and \code{offender_characteristics_relationship_to_victim} respectively, which
each return a data frame.

The function then binds these 5 parts together into a single data frame to get the full "Offender characteristics" table
before processing it into a \code{GPTable} object (see the User Guide for more details) which is what is returned.
}
\examples{
config = autotable::process_config(yaml::read.yaml("configs/offender_characteristics/BURGLARY.yaml"))
master_config = yaml::read_yaml(file = "configs/MASTER.yaml")
offender_characteristics(config = config, master_config = master_config)

}
\seealso{
See the following URL for an example of a Nature of Crime dataset (where the headline crime type in the example is
"Violence"): \url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/natureofcrimetablesviolence}
}
\author{
Eleanor Scott-Allen, Katie Edser, Matthew Morrison, Nicholas O'Neil; ONS Centre for Crime and Justice, Titchfield, PO15 5RR.
}
