% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/offender_characteristics_able_to_say_something.R
\name{offender_characteristics_able_to_say_something}
\alias{offender_characteristics_able_to_say_something}
\title{offender_characteristics_able_to_say_something}
\usage{
offender_characteristics_able_to_say_something(config, master_config, subtable)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
