% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/offender_characteristics_number_of_offenders.R
\name{offender_characteristics_number_of_offenders}
\alias{offender_characteristics_number_of_offenders}
\title{offender_characteristics_number_of_offenders}
\usage{
offender_characteristics_number_of_offenders(config, master_config, subtable)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
