% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/offender_characteristics_relationship_to_victim.R
\name{offender_characteristics_relationship_to_victim}
\alias{offender_characteristics_relationship_to_victim}
\title{offender_characteristics_relationship_to_victim}
\usage{
offender_characteristics_relationship_to_victim(
  config,
  master_config,
  subtable
)
}
\arguments{
\item{config}{To do.}

\item{master_config}{To do.}

\item{subtable}{To do.}
}
\value{
To do.
}
\description{
To do.
}
\details{
To do.
}
\examples{
To do.

}
