% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/proportion_of_offenders_under_the_influence_of_alcohol.R
\name{proportion_of_offenders_under_the_influence_of_alcohol}
\alias{proportion_of_offenders_under_the_influence_of_alcohol}
\title{proportion_of_offenders_under_the_influence_of_alcohol}
\usage{
proportion_of_offenders_under_the_influence_of_alcohol(
  config,
  master_config,
  subtable
)
}
\arguments{
\item{config}{A string corresponding to the path of a YAML file, where this YAML serves as the configuration file for the
function. This YAML must exhibit a particular structure/format, since the function accesses certain levels and fields of it
by name. For more information on the format required, please see the User Guide (found in the "resources" folder of the
project) or inspect a valid YAML file from previous successful runs.

#'@param master_config A string corresponding to a folder path whereat the VF datasets for each year are stored as .csv
files.

#'@param subtable An integer indicating which subtable to consider; if we denote the value of \code{subtable} by \code{n},
then the subtable considered will be that corresponding to the \code{n}th item in
\code{config$OFFENCE_CODES}.}
}
\value{
Returns a dataframe with the requisite percentages and unweighted bases for the POD table
}
\description{
Builds the "Proportion that offender under influence of alcohol" (IoA) sub table for a Nature of Crime dataset.
}
\details{
This function builds the IoA table (specified by \code{subtable}) based largely on a configuration
file specified by \code{config}.
}
\examples{
proportion_of_offenders_under_the_influence_of_alcohol(config = "configs/proportion_of_offenders_being_under_the_influence_of_alcohol/ROBBERY.yaml", master_config =
master_config, subtable = 1)

}
