% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_zeroweighted_rows.R
\name{remove_zeroweighted_rows}
\alias{remove_zeroweighted_rows}
\title{remove_zeroweighted_rows}
\usage{
remove_zeroweighted_rows(df, weightvar)
}
\arguments{
\item{df}{- a file to remove the zero weighted rows from}

\item{weightvar}{- the variable to use as the criterion for whether a row should be deleted}
}
\value{
Data Frame
}
\description{
Strips out rows where the record has a weight of zero
}
\details{
Given a specified weight column, this function deletes records from the input file that have a weight of 0 and returns a data frame with these records removed.
}
\examples{
remove_zeroweighted_rows(mtcars, "vs")

}
