% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing_daylight_or_dark.R
\name{timing_daylight_or_dark}
\alias{timing_daylight_or_dark}
\title{timing_daylight_or_dark}
\usage{
timing_daylight_or_dark(config, master_config, subtable)
}
\arguments{
\item{config}{Documentation to be done later.}

\item{master_config}{Documentation to be done later.}

\item{subtable}{Documentation to be done later.}
}
\value{
Documentation to be done later.
}
\description{
Documentation to be done later.
}
\details{
Documentation to be done later.
}
\examples{
Documentation to be done later.

}
