% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timing_time_of_day.R
\name{timing_time_of_day}
\alias{timing_time_of_day}
\title{timing_time_of_day}
\usage{
timing_time_of_day(config, master_config, subtable)
}
\arguments{
\item{config}{Documentation to be done later.}

\item{master_config}{Documentation to be done later.}

\item{subtable}{Documentation to be done later.}
}
\value{
Documentation to be done later.
}
\description{
Documentation to be done later.
}
\details{
Documentation to be done later.
}
\examples{
Documentation to be done later.

}
