% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_of_damage.R
\name{type_of_damage}
\alias{type_of_damage}
\title{damage type}
\usage{
type_of_damage(config, master_config, return_as_dataframes = FALSE)
}
\arguments{
\item{config}{A string corresponding to the path of a YAML file, where this YAML serves as the configuration file for
the function. This YAML must exhibit a particular structure/format, since the function accesses certain levels and fields by
name. For more information on the format required, please see the User Guide (found in the "resources" folder of the
project) or inspect a valid YAML file from previous successful runs.}

\item{master_config}{Documentation to be done later.}

\item{return_as_dataframes}{A Boolean; \code{TRUE} means that the subtables are returned as a list of dataframes, while
\code{FALSE} (the default) means that they are returned as a list of gptable objects.}
}
\value{
Returns a list of the Table's subtables as either gptable objects (by default) or as R dataframes (should
\code{return_as_dataframes} be set to \code{TRUE}.
}
\description{
Builds the "Emotional impact of incidents" Table for a headline crime type, as seen in the "Nature of Crime"
dataset for that headline crime type. See the following URL for an example of a Nature of Crime dataset, for the headline
crime type "Vehicle-related theft":
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/natureofcrimevehiclerelatedtheft}.
}
\details{
This function builds an "Type of Damage" Table based on a configuration file, where this
configuration file corresponds to one of the headline crime types. Each subtable of the Table is processed in turn,
where the bulk of this processing is handled by \code{build_any_damage_caused} and
\code{build_type_of_damage_caused}, which build the two halves/parts of the subtable respectively. By
default, each subtable is converted to a \code{gptable} object and stored in a list as the final step; if, however,
\code{return_as_dataframes} is set to \code{TRUE}, a list of R dataframes will be returned instead.
}
\examples{
type_of_damage(config = "configs/type_of_damage/BURGLARY.yaml", data_directory = "data/backseries/")

}
