% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/type_of_damage_build_any_damage_caused.R
\name{type_of_damage_build_any_damage_caused}
\alias{type_of_damage_build_any_damage_caused}
\title{type_of_damage_build_any_damage_caused}
\usage{
type_of_damage_build_any_damage_caused(config, master_config, subtable)
}
\arguments{
\item{config}{A string corresponding to the path of a YAML file, where this YAML serves as the configuration file for the
function. This YAML must exhibit a particular structure/format, since the function accesses certain levels and fields of it
by name. For more information on the format required, please see the User Guide (found in the "resources" folder of the
project) or inspect a valid YAML file from previous successful runs.}

\item{master_config}{A string corresponding to a folder path whereat the VF datasets for each year are stored as .csv
files.}

\item{subtable}{An integer indicating which subtable to consider; if we denote the value of \code{subtable} by \code{n},
then the subtable considered will be that corresponding to the \code{n}th item in
\code{config$BOTH_PARTS_OF_THE_TABLE$OFFENCE_CODES}.}
}
\value{
Returns a dataframe with the requisite percentages and unweighted bases for the DC part of the specified subtable.
}
\description{
Builds the "Some form of damage caused" (DC) part/half of one subtable of an " Type of damage caused" Table for a Nature of Crime dataset.
}
\details{
This function builds the DC part of one subtable (specified by \code{subtable}) based largely on a configuration
file specified by \code{config}.
}
\examples{
build_any_damage_caused(config = "configs/type_of_damage/BURGLARY.yaml", data_directory =
"data/backseries/", subtable = 2)

}
