% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/where_incident_occurred.R
\name{where_incident_occurred}
\alias{where_incident_occurred}
\title{where_incident_occurred}
\usage{
where_incident_occurred(config, master_config, return_as_dataframes = FALSE)
}
\arguments{
\item{config}{A string corresponding to the path of a YAML file, where this YAML serves as the configuration file for
the function. This YAML must exhibit a particular structure/format, since the function accesses certain levels and fields by
name. For more information on the format required, please see the User Guide (found in the "resources" folder of the
project) or inspect a valid YAML file from previous successful runs.}

\item{master_config}{A string corresponding to a folder path, whereat the VF datasets for each year are stored as .csv
files.}

\item{return_as_dataframes}{A Boolean; \code{TRUE} means that the subtables are returned as a list of dataframes, while
\code{FALSE} (the default) means that they are returned as a list of gptable objects.}
}
\value{
Returns a list of the Table's subtables as either gptable objects (by default) or as R dataframes (should
\code{return_as_dataframes} be set to \code{TRUE}.
}
\description{
Builds the "Where incidents occurred" Table for a headline crime type, as seen in the "Nature of Crime"
dataset for that headline crime type. See the following URL for an example of a Nature of Crime dataset, for the headline
crime type "Robbery":
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/crimeandjustice/datasets/natureofrobbery}.
}
\details{
This function builds an "Point of entry in incidents" Table based on a configuration file, where this
configuration file corresponds to one of the headline crime types. The table is processed
where the processing is handled by \code{build_where_incident_occurred} By
default, each subtable is converted to a \code{gptable} object and stored in a list as the final step; if, however,
\code{return_as_dataframes} is set to \code{TRUE}, a list of R dataframes will be returned instead.
}
\examples{
where_incident_occurred(config = "configs/where_incident_occurred/ROBBERY.yaml", master_config = master_config)

}
