% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test_crime.R
\name{z_test_crime}
\alias{z_test_crime}
\title{z_test_crime}
\usage{
z_test_crime(
  estimate_1,
  estimate_2,
  base_1,
  base_2,
  design_1 = 1.2,
  design_2 = 1.2,
  threshold = 1.96,
  significance_indicator = 1,
  insignificance_indicator = 0
)
}
\arguments{
\item{estimate_1}{Documentation to be added later.}

\item{estimate_2}{Documentation to be added later.}

\item{base_1}{Documentation to be added later.}

\item{base_2}{Documentation to be added later.}

\item{design_1}{Documentation to be added later.}

\item{design_2}{Documentation to be added later.}

\item{threshold}{Documentation to be added later.}

\item{significance_indicator}{Documentation to be added later.}

\item{insignificance_indicator}{Documentation to be added later.}
}
\value{
Returns the back-series as a data frame of percentages with the unweighted base as the last row.
}
\description{
Documentation to be added later.
}
\details{
Documentation to be added later.
}
\examples{

}
